[MIN-354] Move balancing tests from runtime in liquidatio-pools module
